refactor frontend components path
