--- frontend/src/components/MultiSeclect.tsx
@@ -1,95 +1,0 @@
-import * as React from "react"
-
-import FormControl from "@mui/material/FormControl"
-import MenuItem from "@mui/material/MenuItem"
-import OutlinedInput from "@mui/material/OutlinedInput"
-import Select, { SelectChangeEvent } from "@mui/material/Select"
-import { Theme, useTheme } from "@mui/material/styles"
-
-const ITEM_HEIGHT = 48
-const ITEM_PADDING_TOP = 8
-const MenuProps = {
-  PaperProps: {
-    style: {
-      maxHeight: ITEM_HEIGHT * 4.5 + ITEM_PADDING_TOP,
-      width: 250,
-    },
-  },
-}
-
-const names = [
-  "Oliver Hansen",
-  "Van Henry",
-  "April Tucker",
-  "Ralph Hubbard",
-  "Omar Alexander",
-  "Carlos Abbott",
-  "Miriam Wagner",
-  "Bradley Wilkerson",
-  "Virginia Andrews",
-  "Kelly Snyder",
-]
-
-const getStyles = (
-  name: string,
-  personName: readonly string[],
-  theme: Theme,
-) => {
-  return {
-    fontWeight:
-      personName.indexOf(name) === -1
-        ? theme.typography.fontWeightRegular
-        : theme.typography.fontWeightMedium,
-  }
-}
-
-const MultipleSelectPlaceholder = () => {
-  const theme = useTheme()
-  const [personName, setPersonName] = React.useState<string[]>([])
-
-  const handleChange = (event: SelectChangeEvent<typeof personName>) => {
-    const {
-      target: { value },
-    } = event
-    setPersonName(typeof value === "string" ? value.split(",") : value)
-  }
-
-  return (
-    <div>
-      <FormControl sx={{ width: 300 }}>
-        <Select
-          sx={{ height: 36, width: 272 }}
-          multiple
-          displayEmpty
-          value={personName}
-          onChange={handleChange}
-          input={<OutlinedInput />}
-          renderValue={(selected) => {
-            if (selected.length === 0) {
-              return <em>Select Group</em>
-            }
-
-            return selected.join(", ")
-          }}
-          MenuProps={MenuProps}
-          inputProps={{ "aria-label": "Without label" }}
-        >
-          <MenuItem disabled value="">
-            <em>Placeholder</em>
-          </MenuItem>
-          {names.map((name) => (
-            <MenuItem
-              key={name}
-              value={name}
-              style={getStyles(name, personName, theme)}
-            >
-              {name}
-            </MenuItem>
-          ))}
-        </Select>
-      </FormControl>
-    </div>
-  )
-}
-
-export default MultipleSelectPlaceholder
